chore: Fix jsdom
