+++ src/tabs/scroll-utils.ts
@@ -1,8 +1,18 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { getIsRtl, getScrollInlineStart } from '@cloudscape-design/component-toolkit/internal';
+import { getIsRtl, getScrollInlineStart, isMotionDisabled } from '@cloudscape-design/component-toolkit/internal';
 
-import smoothScroll from './smooth-scroll';
+const smoothScroll = (element: HTMLElement, to: number) => {
+  if (isMotionDisabled(element) || !element.scrollTo) {
+    element.scrollLeft = to;
+    return;
+  }
+  // istanbul ignore next: unit tests always have motion disabled
+  element.scrollTo({
+    left: to,
+    behavior: 'smooth',
+  });
+};
 
 export const onPaginationClick = (
   headerBarRef: React.RefObject<HTMLUListElement>,
